Create a structure to specify data on students given below:
Roll number, Name, Department, Course, Year of joining

Assume that there are not more than 450 students in the
college. 

(a) Write a function to print names of all students who joined in a particular year.
(b) Write a function to print the data of a student whose roll number is given.