Create a structure to specify data of customers in a bank. The data to be stored is Account Number, Name, Balance in the account. Assume a maximum of 200 customers in the bank.
(a) Write a function to print the Account number and name of each customer with a balance below Rs. 1000. 

(b) If a customer request for withdrawal or deposit, it is given in the form:

Acct. no, amount, code (1 for the deposit, 0 for withdrawal)

Write a program to give a message, "The balance is insufficient for the specified withdrawal".